[#123] Minor refactor from PR
